Create code assigner
google app script code so that it looks for the code of column F of Sheet 2 in column B of Sheet 1 and assigns the code of column P of Sheet 2 in column E of Sheet 1. In the case If there are duplicate codes in column B, just assign the code of F to the first of the duplicates.